requirements.txt
langchain==0.0.339
streamlit
openai
pypdf
faiss-cpu
accelerate
python-dotenv
tiktoken
chromadb